add log messages, merge query_unbound_ppns() into find_unbound_pps()

+++ src/EventHandler.py
@@ -50,7 +50,7 @@ class EventHandler:
         self.tasks.append(asyncio.create_task(self.omm_mgr.start_communication(request_lock=request_lock)))
 
         # Collect unbound PPNs task, collects unbound devices in OMM and assigns them temp accounts
-        self.tasks.append(asyncio.create_task(self.query_unbound_ppns()))
+        self.tasks.append(asyncio.create_task(self.find_unbound_pps()))
 
         # Handle Device Registrations
         self.tasks.append(asyncio.create_task(self.handle_device_registrations()))
@@ -111,6 +111,13 @@ class EventHandler:
                 from_user = self.omm_mgr.omm.find_user({'num': temp_number})
                 # find OMM user with the corresponding token
                 to_user = self.omm_mgr.omm.find_user({'hierarchy2': token})
+                if not from_user:
+                    self.logger.warning(f'Failed to fetch temp user (temp_num:{temp_number}) on registration! Can\'t transfer PP!')
+                    continue
+                if not to_user:
+                    self.logger.warning(f'Failed to fetch OMM user (token:{token}) on registration! Can\'t transfer PP!')
+                    continue
+                self.logger.info(f'Transferring PP (ppn:{from_user.ppn}) to OMM user (uid: {to_user.uid}, number: {to_user.num}.)')
                 # transfer PP to real user
                 self.omm_mgr.transfer_pp(int(from_user.uid), int(to_user.uid), int(from_user.ppn))
                 # delete temporary user, both in OMM and Asterisk
@@ -144,7 +151,8 @@ class EventHandler:
 
     def do_sip_extension_update(self, event_data):
         number = event_data['number']
-        password = event_data['password']
+        sip_password = event_data['password']
+        self.logger.info(f'Processing SIP extension update for number {number}.')
 
         # delete DECT extension, if present
         if number in self.omm_mgr.users:
@@ -152,43 +160,53 @@ class EventHandler:
 
         # SIP extension already exists, only a password update is required
         if self.asterisk_mgr.check_for_user(number=number):
-            self.asterisk_mgr.update_password(number=number, new_password=password)
+            self.asterisk_mgr.update_password(number=number, new_password=sip_password)
 
         # new SIP extension
         else:
-            self.asterisk_mgr.create_user(number=number, sip_password=password)
+            self.asterisk_mgr.create_user(number=number, sip_password=sip_password)
 
     def do_dect_extension_update(self, event_data):
         name = event_data['name']
         number = event_data['number']
         token = event_data['token']
+        self.logger.info(f'Processing DECT extension update for number {number}.')
 
         # if user already exists, update user entry
         if number in self.omm_mgr.users:
+            self.logger.info(f'Updating OMM user {number}.')
             self.omm_mgr.update_user_info(number=number, name=name, token=token)
 
         # else, create a new user
         else:
+            self.logger.info(f'Creating new Asterisk and OMM user for number {number}.')
             # make sure that any existing SIP user is being deleted beforehand
             if self.asterisk_mgr.check_for_user(number=number):
+                self.logger.warning('Deleting existing Asterisk user that would clash with the newly created one!')
                 self.asterisk_mgr.delete_user(number=number)
 
-            sip_password = self.asterisk_mgr.create_user(number=number)
+            sip_password = utils.create_password('alphanum', self.all_config['asterisk']['password_length'])
+            self.asterisk_mgr.create_user(number=number, sip_password=sip_password)
             self.omm_mgr.create_user(name=name, number=number, token=token, sip_user=number, sip_password=sip_password)
 
     def do_delete_extension(self, event_data):
         number = event_data['number']
-        self.asterisk_mgr.delete_user(number=number)
+        if self.asterisk_mgr.check_for_user(number):
+            self.logger.info(f'Deleting Asterisk user {number}.')
+            self.asterisk_mgr.delete_user(number=number)
         if number in self.omm_mgr.users:
+            self.logger.info(f'Deleting OMM user {number}.')
             self.omm_mgr.delete_user(number=number)
 
     def do_rename_extension(self, event_data):
         old_number = event_data['old_extension']
         new_number = event_data['new_extension']
         if self.asterisk_mgr.check_for_user(old_number):
+            self.logger.info(f'Renaming extension from {old_number} to {new_number} in Asterisk.')
             self.asterisk_mgr.move_user(old_number, new_number)
 
         if old_number in self.omm_mgr.users:
+            self.logger.info(f'Renaming extension from {old_number} to {new_number} in OMM.')
             self.omm_mgr.move_user(old_number, new_number)
 
     def do_unsubscribe_device(self, event_data):
@@ -196,25 +214,25 @@ class EventHandler:
         # TODO: OMM MAGIC!
         return
 
-    def register_devices(self):
-        for device in self.omm_mgr.omm.get_devices():
-            if device.relType != 'Unbound':
-                continue
-            temp_number = f'010' + utils.create_password('num', self.all_config['asterisk']['temp_num_length'])
-            temp_password = utils.create_password('alphanum', self.all_config['asterisk']['password_length'])
-            while self.asterisk_mgr.check_for_user(temp_number):
-                temp_number = f'010' + utils.create_password('num', self.all_config['asterisk']['temp_num_length'])
-
-            omm_user = self.omm_mgr.create_user(name='Unbound Handset', number=temp_number, sip_user=temp_number,
-                                                sip_password=temp_password)
-            self.omm_mgr.omm.attach_user_device(uid=int(omm_user.uid), ppn=int(device.ppn))
-            self.asterisk_mgr.create_user(number=temp_number, sip_password=temp_password, temporary=True)
-
-    async def query_unbound_ppns(self):
+    async def find_unbound_pps(self):
         try:
             self.logger.info('Now looking for unbound PPs.')
             while True:
-                self.register_devices()
+                for device in self.omm_mgr.omm.get_devices():
+                    if device.relType != 'Unbound':
+                        continue
+                    temp_number = f'010' + utils.create_password('num', self.all_config['asterisk']['temp_num_length'])
+                    temp_password = utils.create_password('alphanum', self.all_config['asterisk']['password_length'])
+                    while self.asterisk_mgr.check_for_user(temp_number):
+                        temp_number = f'010' + utils.create_password('num',
+                                                                     self.all_config['asterisk']['temp_num_length'])
+                    self.logger.info(f'Assigning unbound device ({device.ppn}) to a temporary user ({temp_number})')
+                    omm_user = self.omm_mgr.create_user(name='Unbound Handset', number=temp_number,
+                                                        sip_user=temp_number,
+                                                        sip_password=temp_password)
+                    self.omm_mgr.omm.attach_user_device(uid=int(omm_user.uid), ppn=int(device.ppn))
+                    self.asterisk_mgr.create_user(number=temp_number, sip_password=temp_password, temporary=True)
+
                 await asyncio.sleep(self.own_config['collect_ppns_interval'])
         except asyncio.CancelledError:
             pass
@@ -224,3 +242,5 @@ class EventHandler:
         for task in self.tasks:
             task.cancel()
         self.asterisk_mgr.close()
+        self.logger.info('Shutdown complete, goodbye.')
+
